simplify binding files organization
